+++ remus/proto/testing/UnitTestJobSubmission.cxx
@@ -57,7 +57,7 @@ remus::common::MeshIOType randomMeshTypes()
 {
   remus::common::MeshIOTypeSet allTypes = remus::common::generateAllIOTypes();
 
-  int selected_index = randomInt(0,allTypes.size());
+  int selected_index = randomInt(0, static_cast<int>(allTypes.size()) );
   int index=0;
   typedef remus::common::MeshIOTypeSet::const_iterator cit;
   for(cit i=allTypes.begin(); i != allTypes.end(); ++i, ++index)
